+++ src/cognitect/aws/ec2_metadata_utils.clj
@@ -15,6 +15,7 @@
 
 (def ^:const ec2-metadata-service-override-system-property "com.amazonaws.sdk.ec2MetadataServiceEndpointOverride")
 (def ^:const dynamic-data-root "/latest/dynamic/")
+(def ^:const imds-v2-token-path "/latest/api/token")
 (def ^:const security-credentials-path "/latest/meta-data/iam/security-credentials/")
 (def ^:const instance-identity-document "instance-identity/document")
 
@@ -58,9 +59,9 @@
                 auth-token
                 (assoc "Authorization" auth-token))}))
 
-(defn get-data [uri http-client]
+(defn- get-response-data [request-map http-client]
   (let [response (a/<!! (retry/with-retry
-                          #(http/submit http-client (request-map (URI. uri)))
+                          #(http/submit http-client request-map)
                           (a/promise-chan)
                           retry/default-retriable?
                           retry/default-backoff))]
@@ -68,32 +69,89 @@
     (when (= 200 (:status response))
       (u/bbuf->str (:body response)))))
 
-(defn get-data-at-path [path http-client]
+(defn ^:deprecated get-data [uri http-client]
+  "DEPRECATED use `request-map`, `get-response-data`"
+  (get-response-data (request-map (URI. uri)) http-client))
+
+(defn ^:deprecated get-data-at-path [path http-client]
+  "DEPRECATED use `build-uri`, `request-map`, `get-response-data`"
   (get-data (build-uri (get-host-address) path) http-client))
 
-(defn get-listing [uri http-client]
+(defn ^:deprecated get-listing [uri http-client]
+  "DEPRECATED use `get-listing-from-response`"
   (some-> (get-data uri http-client) str/split-lines))
 
-(defn get-listing-at-path [path http-client]
+(defn ^:deprecated get-listing-at-path [path http-client]
+  "DEPRECATED use `get-listing-from-response`"
   (get-listing (build-uri (get-host-address) path) http-client))
 
-(defn get-ec2-instance-data [http-client]
-  (some-> (build-path dynamic-data-root instance-identity-document)
-          (get-data-at-path http-client)
-          (json/read-str :key-fn keyword)))
+(defn- add-header-when-truthy
+  "Add a header to a request map IFF the header value is truthy, return the 'updated' request map,
+  else return the unmodified request map."
+  [request header value]
+  (if value
+    (assoc-in request [:headers header] value)
+    request))
+
+(defn- get-listing-from-response [data]
+  (some-> data str/split-lines))
+
+(defn get-ec2-instance-data
+  ([http-client]
+   (get-ec2-instance-data http-client nil))
+  ([http-client IMDSv2-token]
+   (some-> (get-host-address)
+           (build-uri (build-path dynamic-data-root instance-identity-document))
+           (URI.)
+           (request-map)
+           (add-header-when-truthy "X-aws-ec2-metadata-token" IMDSv2-token)
+           (get-response-data http-client)
+           (json/read-str :key-fn keyword))))
 
 (defn get-ec2-instance-region
-  [http-client]
-  (:region (get-ec2-instance-data http-client)))
+  ([http-client]
+   (get-ec2-instance-region http-client nil))
+  ([http-client IMDSv2-token]
+   (:region (get-ec2-instance-data http-client IMDSv2-token))))
 
 (defn container-credentials [http-client]
   (let [endpoint (or (when-let [path (u/getenv container-credentials-relative-uri-env-var)]
                        (str (get-host-address) path))
                      (u/getenv container-credentials-full-uri-env-var))]
-    (some-> endpoint (get-data http-client) (json/read-str :key-fn keyword))))
+    (some-> endpoint
+            (URI.)
+            (request-map)
+            (get-response-data http-client)
+            (json/read-str :key-fn keyword))))
 
-(defn instance-credentials [http-client]
-  (when (not (in-container?))
-    (when-let [cred-name (first (get-listing-at-path security-credentials-path http-client))]
-      (some-> (get-data-at-path (str security-credentials-path cred-name) http-client)
-              (json/read-str :key-fn keyword)))))
+(defn instance-credentials
+  ([http-client]
+   (instance-credentials http-client nil))
+  ([http-client IMDSv2-token]
+   (when (not (in-container?))
+     (when-let [cred-name (-> (get-host-address)
+                              (build-uri security-credentials-path)
+                              (URI.)
+                              (request-map)
+                              (add-header-when-truthy "X-aws-ec2-metadata-token" IMDSv2-token)
+                              (get-response-data http-client)
+                              (get-listing-from-response)
+                              first)]
+       (some-> (get-host-address)
+               (build-uri (str security-credentials-path cred-name))
+               (URI.)
+               (request-map)
+               (add-header-when-truthy "X-aws-ec2-metadata-token" IMDSv2-token)
+               (get-response-data http-client)
+               (json/read-str :key-fn keyword))))))
+
+(defn IMDSv2-token
+  "Retrieve and return an IMDS v2 session token, or nil if IMDS v2 is not in effect."
+  [http-client]
+  (-> (get-host-address)
+      (build-uri imds-v2-token-path)
+      (URI.)
+      (request-map)
+      (assoc :request-method :put)
+      (add-header-when-truthy "X-aws-ec2-metadata-token-ttl-seconds" "21600")
+      (get-response-data http-client)))