+++ test/src/cognitect/aws/region_test.clj
@@ -2,16 +2,15 @@
 ;; All rights reserved.
 
 (ns cognitect.aws.region-test
-  (:require [clojure.test :refer [deftest is testing use-fixtures]]
+  (:require [clojure.test :refer [deftest is testing]]
             [clojure.java.io :as io]
             [clojure.core.async :as a]
+            [cognitect.aws.client.shared :as shared]
             [cognitect.aws.region :as region]
             [cognitect.aws.util :as u]
             [cognitect.aws.test.utils :as tu]
             [cognitect.aws.ec2-metadata-utils :as ec2-metadata-utils]
-            [cognitect.aws.ec2-metadata-utils-test :as ec2-metadata-utils-test]))
-
-(use-fixtures :once ec2-metadata-utils-test/test-server)
+            [cognitect.aws.test.ec2-metadata-utils-server :as ec2-metadata-utils-test-server]))
 
 (deftest chain-region-provider-test
   (let [r  "us-east-1"
@@ -43,19 +42,30 @@
 
 (deftest instance-region-provider-test
   (testing "provider caches the fetched value"
-    (let [orig-get-region-fn ec2-metadata-utils/get-ec2-instance-region
-          request-counter    (atom 0)
-          fetch-counter      (atom 0)]
-      (with-redefs [ec2-metadata-utils/get-ec2-instance-region
-                    (fn [http]
-                      (swap! fetch-counter inc)
-                      (orig-get-region-fn http))]
-        (let [num-requests 10
-              p            (region/instance-region-provider ec2-metadata-utils-test/*http-client*)
-              chans        (repeatedly num-requests
-                                       #(do
-                                          (swap! request-counter inc)
-                                          (region/fetch-async p)))]
-          (is (apply = "us-east-1" (map #(a/<!! %) chans)))
-          (is (= num-requests @request-counter))
-          (is (= 1 @fetch-counter)))))) ())
+    (ec2-metadata-utils-test-server/with-test-server
+      (let [orig-get-region-fn ec2-metadata-utils/get-ec2-instance-region
+            request-counter    (atom 0)
+            fetch-counter      (atom 0)]
+        (with-redefs [ec2-metadata-utils/get-ec2-instance-region
+                      (fn [http imdsv2-token]
+                        (swap! fetch-counter inc)
+                        (orig-get-region-fn http imdsv2-token))]
+          (let [num-requests 10
+                p            (region/instance-region-provider (shared/http-client))
+                chans        (repeatedly num-requests
+                                         #(do
+                                            (swap! request-counter inc)
+                                            (region/fetch-async p)))]
+            (is (apply = "us-east-1" (map #(a/<!! %) chans)))
+            (is (= num-requests @request-counter))
+            (is (= 1 @fetch-counter))))))))
+
+(deftest instance-region-provider-test-not-IMDSv2-compliant
+  (testing "returns nil for IMDS v2 server"
+    (ec2-metadata-utils-test-server/with-IMDSv2-test-server
+      (is (nil? (region/fetch (region/instance-region-provider (shared/http-client))))))))
+
+(deftest instance-region-IMDS-v2-provider-test
+  (testing "provider for IMDS v2 server"
+    (ec2-metadata-utils-test-server/with-IMDSv2-test-server
+      (is (= "us-east-1" (region/fetch (region/instance-region-IMDS-v2-provider (shared/http-client))))))))